Update DL 06 - Horovod.py
fetch_california_housing returns a dictionary with data as an ndarray and target as an array
if the argument as_frame=True was passed, it would return it as a dataframe. 

Horovod doesn't work with pandas right? Only numpy or parquet via petastorm

I am still learning about this so I could be horribly wrong.